pyproject.toml
[project]
name = "deo-llm-reframing"
version = "1.0.0"
description = "Testing Distance-Engagement Oscillation as a prompting framework for creative reframing in LLMs"
readme = "README.md"
requires-python = ">=3.12"
license = "CC-BY-4.0"
authors = [{ name = "Gokhan Gokmen", email = "go.gkmn@gmail.com" }]
dependencies = [
    "anthropic>=0.40.0",
    "matplotlib>=3.10.8",
    "numpy>=2.4.3",
    "openai>=2.29.0",
    "openpyxl>=3.1.5",
    "python-dotenv>=1.0.0",
    "scipy>=1.17.1",
]
